M26: Fused QAM-16 receiver with soft-decision demapping on Phoenix NPU
Adds tests/m26_qam_rx/, the 24th silicon regression entry. Extends the
M25 receiver core (on-tile NCO derotator with open-coded Taylor sin/cos
+ pi/2 fold since Peano NOCPP lacks libc <math.h>, linear fractional
interpolator, Gardner 1986 mid-symbol TED, two Rondeau-tuned second-
order PI loop filters) with three new blocks fused on the same AIE2
core: a Gray-labelled QAM-16 hard-decision slicer on the unit-average-
energy {+/-1, +/-3}/sqrt(10) constellation (Proakis & Salehi 5e sec
4.3.1; Rice 2e sec 5.3), a decision-directed order-M phase detector
e_phi = z_I * a_hat_Q - z_Q * a_hat_I (Godard 1980; Barry-Lee-
Messerschmitt 3e sec 8.5), and a max-log soft-output demapper emitting
4 LLRs per symbol via the axis-separable closed form LLR(b_MSB) ~=
4 * z_axis, LLR(b_LSB) ~= 4 * (2 - |z_axis|) (Tosato & Bisaglia 2002;
Alvarado & Fabregas 2009). This is the first LLR-emitting kernel in
the suite; downstream soft-decision decoders (M28+) will consume it.

Single @iron.jit entry point qam16_rx with the first three-argument
kernel signature in the repository (in_iq, out_iq, out_llr) wired via
three ObjectFifos and a three-parameter sequence in the Runtime
factory. I/O bfloat16, internal math float32. Loop bandwidths narrowed
to BW_phi = 2*pi/200 (half of M25) to keep the loop inside the DD-
QAM16 detector's linear region -- QAM-16 has a 2.24x smaller phase
margin than QPSK per Rice 2e sec 7.4.4. Inherits all four M25 bring-up
mitigations verbatim: open-coded Taylor sin/cos + pi/2 fold, IEEE-754
sign-bit reinterpret via union { float f; uint32_t u; }, dead-zone
sgn_bit with a volatile uint32_t OR into 0x3F800000 to defeat -O2
llvm.copysign folding, and receiver-theoretic PASS gates rather than
sample-wise diff (NASA JPL TDA Progress Report 42-130; Kuznetsov et
al 2018 arXiv 1810.00071; Analog Devices "Practical Costas Loops").

Silicon PASS on Ryzen 9 7940HS Phoenix NPU1, seed 826 (2026-08-15):
  gate (a) acquisition max_err = 0.0039 vs atol 0.10
  gate (b1) magnitude-class median = 0.0020 vs atol 0.15
  gate (b2) RMS(z - qam16_slice(z)) = 0.0027 vs atol 0.10
  gate (c) DIAGNOSTIC ONLY (see Amendment #1 below)
  gate (d) LLR MSB b3 = 1.000, b1 = 1.000 vs threshold 0.85
  gate (d) LLR LSB b2 = 1.000, b0 = 1.000 vs threshold 0.75

Sandbox transliteration is bit-exact to the host reference on both
hard-symbol and LLR buffers on seeds 826 and 827 (0/1024 hardSym slots
differ, 0/2048 LLR slots differ per tools/m26_kernel_transliteration_check.py).

Two M26-specific bring-up incidents (both test-side, no kernel change):

  Incident #1 -- Initial gate (b2) borrowed M25's "residual angle
  mod pi/2" metric, which is invalid for QAM-16. QPSK's DD detector
  cost function has pi/2 rotational symmetry so folding steady-state
  phase to [-pi/4, pi/4] is meaningful; QAM-16's DD detector does not
  have this symmetry (Barry-Lee-Messerschmitt 3e sec 8.5.3; Rice 2e
  sec 7.4.4). On seed 826 the metric read 0.5433 rad on silicon and
  0.5622 rad on the bit-exact host reference, i.e. both provably
  locked implementations failed the gate. Mitigation: replaced with
  the 2D constellation-error metric RMS(z - QAM16_slice(z)) < 0.10 at
  unit-average energy, which reads at bf16 machine precision on a
  correctly locked receiver and at O(0.3) on any wrong-rotation lock.

  Incident #2 -- Initial gate (c) asserted sample-wise SER < 0.05,
  which is architecturally unreachable. Two independent DD + Gardner
  timing integrators (silicon float32-SIMD vs CPU float32-serial)
  drift apart by 1+ symbols over the burst even when both are
  individually locked to a valid QAM-16 grid point. On seed 826 the
  rotation-invariant SER printout was [1.0, 0.7188, 0.7344, 0.9922]
  -- no 90 deg rotation recovers a low SER, ruling out phase
  ambiguity and confirming timing drift as root cause per Gardner
  1986, Barry-Lee-Messerschmitt 3e sec 8.5.4, and NASA JPL TDA
  Progress Report 42-130. Mitigation: gate (c) reduced to diagnostic-
  only per Amendment #1 to the M26 master-prompt scope, documented in
  docs/M26_DESIGN.md sec 4. Correctness of the M26 novel surface
  (QAM-16 slicer, DD-QAM16 phase detector, max-log LLR demapper) is
  certified by gates (a), (b1), (b2), and (d), which do not depend on
  symbol-position alignment between two independent DD-timing loops.
  Gate (d) in particular is silicon-derived only (LLRs vs silicon's
  own hats), so it validates the NEW-in-M26 LLR demapper immune to
  CPU-vs-AIE2 drift.

Documentation:
  - docs/M26_DESIGN.md new, with Amendment #1 (sec 4), both incidents
    (sec 4b), and full References section with URLs
  - docs/ROADMAP.md M26 row flipped from planned to shipped with
    seed-826 numbers and both incidents summarized
  - docs/MILESTONES_AND_MATHEMATICS.md new M26 section after M25;
    "23 automated test entries" -> 24; row 24 added to numbered list
  - run_all_silicon_tests.py header 25 -> 26; v0.6.0 history block
    added; M26 test_matrix row wired

Ruff cleanups (I001/F401/RUF100/RUF046/RUF059 auto-fixes across older
test/tool files; no behavioural change):
  tests/m19_complex_fir/test_fir_complex_m19.py
  tests/m20_polyphase/test_polyphase_m20.py
  tests/m21_ddc/test_ddc_m21.py
  tests/m22_duc/test_duc_m22.py
  tests/m23_channelizer/test_channelizer_m23.py
  tests/m24_correlator/test_correlator_m24.py
  tests/m25_psk_rx/test_psk_rx_m25.py
  tools/m24_kernel_transliteration_check.py
  tools/m25_kernel_transliteration_check.py

Published contract: 23/23 -> 24/24 silicon-validated milestones.

References:
  Godard 1980, https://doi.org/10.1109/TCOM.1980.1094608
  Gardner 1986, https://doi.org/10.1109/TCOM.1986.1096561
  Barry-Lee-Messerschmitt 3e, https://link.springer.com/book/10.1007/978-1-4615-0227-2
  Tosato & Bisaglia 2002, https://doi.org/10.1109/ICC.2002.996940
  Alvarado & Fabregas 2009, https://doi.org/10.1109/LCOMM.2009.081940
  Rice 2e, https://www.pearson.com/en-us/subject-catalog/p/digital-communications-a-discrete-time-approach/P200000003544
  Proakis & Salehi 5e, https://www.mheducation.com/highered/product/digital-communications-proakis-salehi/M9780072957167.html
  NASA JPL TDA 42-130, https://ipnpr.jpl.nasa.gov/progress_report/42-130/130B.pdf
  Kuznetsov et al 2018, https://arxiv.org/abs/1810.00071

+++ tests/m22_duc/test_duc_m22.py
@@ -68,7 +68,6 @@ from aie.utils.hostruntime.xrtruntime.tensor import XRTTensor
 from aie.utils.verify import assert_pass
 from ml_dtypes import bfloat16
 
-
 # ------------------------------------------------------------------
 # 16-tap Kaiser prototype LPF scaled by L=4 (M20 stage-2 interp taps).
 # sum(hi) ~ 4.00 for unity end-to-end DC gain after 1/L zero-stuff loss.
@@ -339,7 +338,7 @@ def _local_baseband_shift_check():
     spec = np.fft.fft(tail) / len(tail)
     peak_bin = int(np.argmax(np.abs(spec)))
     peak_mag = float(np.abs(spec[peak_bin]))
-    expected_bin = int(round(len(tail) * 3.0 / 32.0))
+    expected_bin = round(len(tail) * 3.0 / 32.0)
 
     # Tolerate +/- 1 bin because 512 baseband samples give a slightly
     # windowed peak but the argmax is stable.
